+++ server/package.json
@@ -34,6 +34,7 @@
     "@routes": "./src/routes",
     "@utils": "./src/utils",
     "@services": "./src/services",
+    "@jobs": "./src/jobs",
     "@events": "./src/events"
   },
   "license": "MIT",
